# CRM-21108 - Greatly improve performance of adding contacts