@MegaphoneJon 2017-08-28
Test failure is related.  I ran CRM_Utils_Token and CRM_Contact_BAO tests, but not the whole test suite because of the overhead (which hopefully this PR will reduce!).  I'll take a look and deal with this.