@totten 2017-08-28
I haven't dug into the patch, but a general comment -- that looks like a really good investigation into why bulk-addition becomes slow.