@monishdeb 2017-08-29
Here's the time difference on running provided php script ```contactprofiler.php```, before and after applying the patch:

1. BEFORE
Start Time: 22:37:28
End Time: 22:37:38
Difference - 10sec

2. AFTER
Start Time: 22:38:10
End Time: 22:38:14
Difference - 4sec 

Also, checked the underlying code and tested the patch on my local. Working fine, didn't cause any unintended regression.